feat: remove unsued filesection and add delete btn to file actions
